CH-444/CH-630 changed uls names
jira[change url names according to the naming convention] https://cozyhometeam.atlassian.net/browse/CH-630?atlOrigin=eyJpIjoiYTMwZmM4ZjI5M2NhNDZmYWJjZDllZTliMzM2NDg5NGIiLCJwIjoiaiJ9
